requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
google-generativeai
gspread
google-auth-oauthlib
pillow
line-bot-sdk
anyio
python-dotenv